Refactored model-usage plugin into domain/presentation modules
