+++ .config/opencode/plugins/model-usage/usage.test.ts
@@ -1,25 +1,8 @@
-import { describe, it, expect, spyOn, mock, afterEach, beforeAll } from "bun:test"
+import { describe, it, expect, spyOn, afterEach } from "bun:test"
 import type { RawUsageRow } from "./db"
 import * as fs from "node:fs"
-
-// ─── Module-level state ──────────────────────────────────────────────────────────
-// We import command.tsx dynamically so we can mock its JSX dependencies first.
-let mod: Awaited<typeof import("./command")>
-
-beforeAll(async () => {
-  // Mock the JSX / UI dependencies that command.tsx imports before loading it
-  mock.module("@opentui/solid", () => ({}))
-  mock.module("@opentui/solid/jsx-dev-runtime", () => ({ jsxDEV: () => null }))
-  mock.module("@opentui/solid/jsx-runtime", () => ({ jsx: () => null, jsxs: () => null, Fragment: () => null }))
-  mock.module("@opencode-ai/plugin/tui", () => ({}))
-  mock.module("solid-js", () => ({
-    onMount: (fn: Function) => {},
-    onCleanup: (fn: Function) => {},
-    createSignal: () => { const fn = (() => {}); return [fn, fn] },
-  }))
-
-  mod = await import("./command")
-})
+import { migrateV2Cache, CACHE_VERSION, MS_PER_DAY, scheduleDiskSave, flushDiskSave, updateMonthCache } from "./cache"
+import { computeUsageDataFromRows, buildHierarchy } from "./usage-domain"
 
 // ─── Helpers ─────────────────────────────────────────────────────────────────────
 
@@ -54,7 +37,7 @@ describe("migrateV2Cache", () => {
         },
       },
     }
-    const result = mod.migrateV2Cache(input)
+    const result = migrateV2Cache(input)
     const model = result.months["123"].models![0]
     expect(model.providerID).toBe("openai")
     expect(model.modelID).toBe("gpt-4")
@@ -76,7 +59,7 @@ describe("migrateV2Cache", () => {
         },
       },
     }
-    const result = mod.migrateV2Cache(input)
+    const result = migrateV2Cache(input)
     const model = result.months["123"].weeks![0].models![0]
     expect(model.providerID).toBe("anthropic")
     expect(model.modelID).toBe("claude-3")
@@ -97,7 +80,7 @@ describe("migrateV2Cache", () => {
         },
       },
     }
-    const result = mod.migrateV2Cache(input)
+    const result = migrateV2Cache(input)
     const model = result.months["123"].days![0].models![0]
     expect(model.providerID).toBe("google")
     expect(model.modelID).toBe("gemini")
@@ -122,7 +105,7 @@ describe("migrateV2Cache", () => {
         },
       },
     }
-    const result = mod.migrateV2Cache(input)
+    const result = migrateV2Cache(input)
     const model = result.months["123"].weeks![0].days![0].models![0]
     expect(model.providerID).toBe("meta")
     expect(model.modelID).toBe("llama")
@@ -139,7 +122,7 @@ describe("migrateV2Cache", () => {
         },
       },
     }
-    const result = mod.migrateV2Cache(input)
+    const result = migrateV2Cache(input)
     const model = result.months["123"].models![0]
     expect(model.providerID).toBe("openai")
     expect(model.modelID).toBe("gpt-4")
@@ -154,8 +137,8 @@ describe("migrateV2Cache", () => {
         "789": {},
       },
     }
-    expect(() => mod.migrateV2Cache(input)).not.toThrow()
-    const result = mod.migrateV2Cache(input)
+    expect(() => migrateV2Cache(input)).not.toThrow()
+    const result = migrateV2Cache(input)
     expect(result.months["123"].models).toBeNull()
     expect(result.months["456"].models).toBeUndefined()
     expect(result.months["789"].models).toBeUndefined()
@@ -163,8 +146,8 @@ describe("migrateV2Cache", () => {
 
   it("sets version to CACHE_VERSION", () => {
     const input = { months: {} }
-    const result = mod.migrateV2Cache(input)
-    expect(result.version).toBe(mod.CACHE_VERSION)
+    const result = migrateV2Cache(input)
+    expect(result.version).toBe(CACHE_VERSION)
   })
 })
 
@@ -182,7 +165,7 @@ describe("buildHierarchy", () => {
         output_tokens: 100,
       }),
     ]
-    const result = mod.buildHierarchy(rows, REF_MONTH_START, REF_MONTH_END)
+    const result = buildHierarchy(rows, REF_MONTH_START, REF_MONTH_END)
     expect(result.models.length).toBeGreaterThan(0)
     expect(result.models[0].providerID).toBe("openai")
     expect(result.models[0].modelID).toBe("gpt-4")
@@ -201,7 +184,7 @@ describe("buildHierarchy", () => {
         output_tokens: 150,
       }),
     ]
-    const result = mod.buildHierarchy(rows, REF_MONTH_START, REF_MONTH_END)
+    const result = buildHierarchy(rows, REF_MONTH_START, REF_MONTH_END)
     expect(result.weeks.length).toBeGreaterThan(0)
     const weekWithData = result.weeks.find(w => w.models !== null && w.models.length > 0)
     expect(weekWithData).toBeDefined()
@@ -220,7 +203,7 @@ describe("buildHierarchy", () => {
         output_tokens: 25,
       }),
     ]
-    const result = mod.buildHierarchy(rows, REF_MONTH_START, REF_MONTH_END)
+    const result = buildHierarchy(rows, REF_MONTH_START, REF_MONTH_END)
     expect(result.days.length).toBeGreaterThan(0)
     const dayWithData = result.days.find(d => d.models !== null && d.models.length > 0)
     expect(dayWithData).toBeDefined()
@@ -230,10 +213,10 @@ describe("buildHierarchy", () => {
 
   it("no providerId/modelId old-format keys appear anywhere in model objects", () => {
     const rows: RawUsageRow[] = [
-      makeRow({ time_created: REF_MONTH_START + mod.MS_PER_DAY * 1, provider_id: "a", model_id: "m1" }),
-      makeRow({ time_created: REF_MONTH_START + mod.MS_PER_DAY * 2, provider_id: "b", model_id: "m2" }),
+      makeRow({ time_created: REF_MONTH_START + MS_PER_DAY * 1, provider_id: "a", model_id: "m1" }),
+      makeRow({ time_created: REF_MONTH_START + MS_PER_DAY * 2, provider_id: "b", model_id: "m2" }),
     ]
-    const result = mod.buildHierarchy(rows, REF_MONTH_START, REF_MONTH_END)
+    const result = buildHierarchy(rows, REF_MONTH_START, REF_MONTH_END)
 
     // Recursively collect all ModelUsage-like objects and check them
     function collectModels(obj: any): any[] {
@@ -264,7 +247,7 @@ describe("buildHierarchy", () => {
   })
 
   it("empty rows produce valid structure with correct day count", () => {
-    const result = mod.buildHierarchy([], REF_MONTH_START, REF_MONTH_END)
+    const result = buildHierarchy([], REF_MONTH_START, REF_MONTH_END)
     expect(result.startMs).toBe(REF_MONTH_START)
     expect(result.endMs).toBe(REF_MONTH_END)
     expect(result.inputTokens).toBe(0)
@@ -285,15 +268,15 @@ describe("buildHierarchy", () => {
 describe("Debounced save", () => {
   afterEach(() => {
     // Clear any leftover timer to avoid cross-test pollution
-    mod.flushDiskSave()
+    flushDiskSave()
   })
 
   it("scheduleDiskSave debounces (only one setTimeout for two calls)", () => {
     const setTimeoutSpy = spyOn(globalThis, "setTimeout").mockImplementation(() => 42 as any)
     const writeSpy = spyOn(fs, "writeFileSync").mockImplementation(() => {})
 
-    mod.scheduleDiskSave()
-    mod.scheduleDiskSave()
+    scheduleDiskSave()
+    scheduleDiskSave()
 
     expect(setTimeoutSpy).toHaveBeenCalledTimes(1)
     expect(writeSpy).toHaveBeenCalledTimes(0)
@@ -306,8 +289,8 @@ describe("Debounced save", () => {
     const setTimeoutSpy = spyOn(globalThis, "setTimeout").mockImplementation(() => 42 as any)
     const writeSpy = spyOn(fs, "writeFileSync").mockImplementation(() => {})
 
-    mod.scheduleDiskSave()
-    mod.flushDiskSave()
+    scheduleDiskSave()
+    flushDiskSave()
 
     expect(writeSpy).toHaveBeenCalledTimes(1)
 
@@ -320,8 +303,8 @@ describe("Debounced save", () => {
     const clearTimeoutSpy = spyOn(globalThis, "clearTimeout").mockImplementation(() => {})
     const writeSpy = spyOn(fs, "writeFileSync").mockImplementation(() => {})
 
-    mod.scheduleDiskSave()
-    mod.flushDiskSave()
+    scheduleDiskSave()
+    flushDiskSave()
 
     expect(clearTimeoutSpy).toHaveBeenCalledWith(42)
     expect(writeSpy).toHaveBeenCalledTimes(1)
@@ -341,7 +324,7 @@ describe("computeUsageDataFromRows", () => {
       makeRow({ provider_id: "openai", model_id: "gpt-4", input_tokens: 200, output_tokens: 100, cost: 0.02 }),
       makeRow({ provider_id: "anthropic", model_id: "claude-3", input_tokens: 300, output_tokens: 150, cost: 0.03 }),
     ]
-    const result = mod.computeUsageDataFromRows(rows)
+    const result = computeUsageDataFromRows(rows)
     expect(result.models).toHaveLength(2)
     const openai = result.models.find(m => m.providerID === "openai" && m.modelID === "gpt-4")!
     expect(openai.totalInput).toBe(300)
@@ -361,7 +344,7 @@ describe("computeUsageDataFromRows", () => {
       makeRow({ provider_id: null, model_id: "gpt-4", input_tokens: 100, output_tokens: 50 }),
       makeRow({ provider_id: "openai", model_id: "gpt-4", input_tokens: 200, output_tokens: 100 }),
     ]
-    const result = mod.computeUsageDataFromRows(rows)
+    const result = computeUsageDataFromRows(rows)
     expect(result.models).toHaveLength(1)
     expect(result.models[0].totalInput).toBe(200)
     expect(result.models[0].totalOutput).toBe(100)
@@ -374,7 +357,7 @@ describe("computeUsageDataFromRows", () => {
       makeRow({ model_id: null, provider_id: "openai", input_tokens: 100, output_tokens: 50 }),
       makeRow({ model_id: "gpt-4", provider_id: "openai", input_tokens: 200, output_tokens: 100 }),
     ]
-    const result = mod.computeUsageDataFromRows(rows)
+    const result = computeUsageDataFromRows(rows)
     expect(result.models).toHaveLength(1)
     expect(result.models[0].totalInput).toBe(200)
     expect(result.models[0].totalOutput).toBe(100)
@@ -388,7 +371,7 @@ describe("computeUsageDataFromRows", () => {
       makeRow({ provider_id: "b", model_id: "m2", input_tokens: 300, output_tokens: 200 }),
       makeRow({ provider_id: "c", model_id: "m3", input_tokens: 50, output_tokens: 25 }),
     ]
-    const result = mod.computeUsageDataFromRows(rows)
+    const result = computeUsageDataFromRows(rows)
     expect(result.models).toHaveLength(3)
     const totals = result.models.map(m => m.totalInput + m.totalOutput)
     expect(totals[0]).toBe(500)
@@ -400,12 +383,12 @@ describe("computeUsageDataFromRows", () => {
     const rows = Array.from({ length: 12 }, (_, i) =>
       makeRow({ provider_id: `p${i}`, model_id: `m${i}` })
     )
-    const result = mod.computeUsageDataFromRows(rows)
+    const result = computeUsageDataFromRows(rows)
     expect(result.models).toHaveLength(10)
   })
 
   it("handles empty input", () => {
-    const result = mod.computeUsageDataFromRows([])
+    const result = computeUsageDataFromRows([])
     expect(result.models).toEqual([])
     expect(result.totalInput).toBe(0)
     expect(result.totalOutput).toBe(0)
@@ -416,7 +399,7 @@ describe("computeUsageDataFromRows", () => {
     const rows = [
       makeRow({ provider_id: "openai", model_id: "gpt-4", input_tokens: -50, output_tokens: -25, cost: -0.01 }),
     ]
-    const result = mod.computeUsageDataFromRows(rows)
+    const result = computeUsageDataFromRows(rows)
     expect(result.models).toHaveLength(1)
     expect(result.models[0].totalInput).toBe(0)
     expect(result.models[0].totalOutput).toBe(0)
@@ -457,8 +440,8 @@ describe("updateMonthCache", () => {
       days: null,
       models: null,
     }
-    mod.updateMonthCache(periodJan)
-    mod.updateMonthCache(periodFeb)
+    updateMonthCache(periodJan)
+    updateMonthCache(periodFeb)
     expect(periodFeb.change).toBe(50) // (1500-1000)/1000*100 = 50
   })
 
@@ -475,7 +458,7 @@ describe("updateMonthCache", () => {
       days: null,
       models: null,
     }
-    mod.updateMonthCache(period)
+    updateMonthCache(period)
     expect(period.change).toBeNull()
   })
 
@@ -528,8 +511,8 @@ describe("updateMonthCache", () => {
       days: [],
       models: [],
     }
-    mod.updateMonthCache(prevPeriod)
-    mod.updateMonthCache(currPeriod)
+    updateMonthCache(prevPeriod)
+    updateMonthCache(currPeriod)
     expect(currFirstWeek.change).toBe(50) // (1350-900)/900*100 = 50
   })
 })
@@ -548,7 +531,7 @@ describe("migrateV2Cache edge cases", () => {
         },
       },
     }
-    const result = mod.migrateV2Cache(input)
+    const result = migrateV2Cache(input)
     const models = result.months["1"].models!
     expect(models).toHaveLength(2)
     for (const m of models) {
@@ -564,8 +547,8 @@ describe("migrateV2Cache edge cases", () => {
   })
 
   it("handles empty months object", () => {
-    const result = mod.migrateV2Cache({ months: {} })
-    expect(result.version).toBe(mod.CACHE_VERSION)
+    const result = migrateV2Cache({ months: {} })
+    expect(result.version).toBe(CACHE_VERSION)
     expect(result.months).toEqual({})
   })
 
@@ -583,8 +566,8 @@ describe("migrateV2Cache edge cases", () => {
         },
       },
     }
-    expect(() => mod.migrateV2Cache(input)).not.toThrow()
-    const result = mod.migrateV2Cache(input)
+    expect(() => migrateV2Cache(input)).not.toThrow()
+    const result = migrateV2Cache(input)
     const model = result.months["1"].weeks![0].models![0]
     expect(model.providerID).toBe("openai")
     expect(model.modelID).toBe("gpt-4")
@@ -600,7 +583,7 @@ describe("buildHierarchy edge cases", () => {
       makeRow({ time_created: REF_MONTH_START }),
       makeRow({ time_created: REF_MONTH_END - 1 }),
     ]
-    const result = mod.buildHierarchy(rows, REF_MONTH_START, REF_MONTH_END)
+    const result = buildHierarchy(rows, REF_MONTH_START, REF_MONTH_END)
     expect(result.days).toHaveLength(31)
     const daysWithData = result.days.filter(d => d.inputTokens > 0)
     expect(daysWithData).toHaveLength(2)
@@ -614,7 +597,7 @@ describe("buildHierarchy edge cases", () => {
 
   it("single row produces correct structure with rest days as zeros", () => {
     const rows = [makeRow({ time_created: REF_MID_MONTH })]
-    const result = mod.buildHierarchy(rows, REF_MONTH_START, REF_MONTH_END)
+    const result = buildHierarchy(rows, REF_MONTH_START, REF_MONTH_END)
     expect(result.days).toHaveLength(31)
 
     const dayWithData = result.days.find(d => d.inputTokens > 0)!
